Rename import_test to test_course since it is doing much more
